Move Supernovae into Events source directory and update related imports, tests, docs, and notebook references.
