+++ docs/slsim.Sources.Events.Supernovae.rst
@@ -1,0 +1,53 @@
+slsim.Sources.Events.Supernovae package
+======================================
+
+Submodules
+----------
+
+slsim.Sources.Events.Supernovae.random\_supernovae module
+--------------------------------------------------
+
+.. automodule:: slsim.Sources.Events.Supernovae.random_supernovae
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+slsim.Sources.Events.Supernovae.supernova module
+-----------------------------------------
+
+.. automodule:: slsim.Sources.Events.Supernovae.supernova
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+slsim.Sources.Events.Supernovae.supernovae\_host\_match module
+-------------------------------------------------------
+
+.. automodule:: slsim.Sources.Events.Supernovae.supernovae_host_match
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+slsim.Sources.Events.Supernovae.supernovae\_lightcone module
+-----------------------------------------------------
+
+.. automodule:: slsim.Sources.Events.Supernovae.supernovae_lightcone
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+slsim.Sources.Events.Supernovae.supernovae\_pop module
+-----------------------------------------------
+
+.. automodule:: slsim.Sources.Events.Supernovae.supernovae_pop
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: slsim.Sources.Events.Supernovae
+   :members:
+   :undoc-members:
+   :show-inheritance:
