+++ docs/slsim.Sources.Events.rst
@@ -1,0 +1,18 @@
+slsim.Sources.Events package
+============================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   slsim.Sources.Events.Supernovae
+
+Module contents
+---------------
+
+.. automodule:: slsim.Sources.Events
+   :members:
+   :undoc-members:
+   :show-inheritance:
