--- docs/slsim.Sources.Supernovae.rst
@@ -1,53 +1,0 @@
-slsim.Sources.Supernovae package
-======================================
-
-Submodules
-----------
-
-slsim.Sources.Supernovae.random\_supernovae module
---------------------------------------------------
-
-.. automodule:: slsim.Sources.Supernovae.random_supernovae
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-slsim.Sources.Supernovae.supernova module
------------------------------------------
-
-.. automodule:: slsim.Sources.Supernovae.supernova
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-slsim.Sources.Supernovae.supernovae\_host\_match module
--------------------------------------------------------
-
-.. automodule:: slsim.Sources.Supernovae.supernovae_host_match
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-slsim.Sources.Supernovae.supernovae\_lightcone module
------------------------------------------------------
-
-.. automodule:: slsim.Sources.Supernovae.supernovae_lightcone
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-slsim.Sources.Supernovae.supernovae\_pop module
------------------------------------------------
-
-.. automodule:: slsim.Sources.Supernovae.supernovae_pop
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: slsim.Sources.Supernovae
-   :members:
-   :undoc-members:
-   :show-inheritance:
