+++ slsim/Sources/Events/Supernovae/supernova.py
