+++ slsim/Sources/SourceCatalogues/SupernovaeCatalog/supernovae_sample.py
@@ -1,8 +1,8 @@
 from astropy.table import Table, hstack
-from slsim.Sources.Supernovae import random_supernovae
-from slsim.Sources.Supernovae.supernovae_lightcone import SNeLightcone
+from slsim.Sources.Events.Supernovae import random_supernovae
+from slsim.Sources.Events.Supernovae.supernovae_lightcone import SNeLightcone
 from slsim.Sources.SourceCatalogues.skypy_galaxy_catalog import GalaxyCatalog
-from slsim.Sources.Supernovae.supernovae_host_match import SupernovaeHostMatch
+from slsim.Sources.Events.Supernovae.supernovae_host_match import SupernovaeHostMatch
 import numpy as np
 from astropy import units
 from scipy import stats
