+++ tests/test_Sources/test_Events/test_Supernovae/test_supernovae_catalog.py
